Adjust Rake tasks for beaker testing
Adjust Rake tasks for testing beaker-rspec suite, beaker suite,
and both in combination.

The beaker hosts config file is set via the BEAKER_setfile
environmental variable. The hosts defined within this file
supersede the host past in as an argument, but the argument is
still required. For an arbitrary hosts config file this provides
the following options.

1. Run the beaker-rspec test suite
    BEAKER_setfile=/path/to/my_hosts.cfg rake beaker:rspec:test[foo,pe]

2. Run the beaker test suite
    BEAKER_setfile=/path/to/my_hosts.cfg rake beaker:test[foo,pe]

3. Run both beaker-rspec and beaker test suites
    BEAKER_setfile=/path/to/my_hosts.cfg rake beaker:test:all[foo,pe]