(MODULES-428) Remove dummy warning
Puppet started raising a warning about multiple default providers due to
a change in puppet. The dummy provider exists to raise an error when
`provider` is not specified because there is no way to declare
`provider` as a required attribute in the type. Even passing a
`provider` value did not get rid of the "multiple default providers
found" warning however.

This commit causes the dummy provider to be the defacto default for all
resources which do not have an explicit provider declared, as the posix
feature is available on basically every operating system on which
vcsrepo works. (There is no way to create an "always default" provider.)