Don't git fetch if sha revision is up-to-date
If you're using a sha as your revision parameter and the current HEAD is
already pointing at that sha, there's no need to do a git fetch.

Making this change should significantly reduce the number of calls made
to the remote Git server; since the catalog hasn't changed, you'll be
up-to-date and not do the fetch call.