feature(git): add keep local changes option
Add an option to keep local changes when revision changes.
As we do a `git checkout --force` all local modifications are removed.
We just add an option that stash the modification before the checkout and
unstash the modification after the checkout.

This change is backward compatible, as by default, we keep the current
behaviour and do not keep local modification.