feat: do not chown excluded files
When using puppetlabs-vcsrepo with the owner parameter (or
with the group parameter), the entire git repository is chowned
recursively on every update.

This chown can be very long if there are a lot of files in the
repository destination path. This commit excludes the files defined in
the `excludes` parameter in the chown.

refs https://tickets.puppetlabs.com/browse/MODULES-5922