Always run as given user, even if identity set
Previously, if the resource specified `user`, then git would be executed
as that user... except if you also provided `identity` to control which
SSH key to use, in which case the code did the setup required for the
git helper script, then called git directly (bypassing the bit that ran
it as the desired user).

This changes it by wrapping up the execution of git in a new method
which contains the logic to run as the specified user if they asked for
that, and uses that wrapper from both places in git_with_identity.