user & path updates in repo cloning acceptance tests
I think this remaining test failure is a confusion over which user it's
SSHing as vs which user had the host keys written to, etc.

I think the run_shell() commands which write out a .ssh/config with
StrictHostKeyChecking can be removed from the tests here too, since the
host key *should* be right, but let's see first if this gets the test to
pass before looking at that.

(If this is wrong, feel free to revert!)