Repeat test user setup here
Repeat the setup of the testssh-user here for this test too.

Copying and pasting this feels very wrong; it should probably be
refactored out into a method that is called by both of the "check out
over SSH" tests, or the two tests turned into subtests in the same
context?